requirements.txt
pyyaml==6.0
eth_account==0.5.7
web3==5.28.0